Move CurrentUserGuild to the models crate
